[MOV] pivot: move pivot table file
This commit moves file
`src/helpers/pivot/spreadsheet_pivot/table_spreadsheet_pivot.ts`
out of `spreadsheet_pivot` which is where all spreadsheet pivot
specific things go.

`SpreadsheetPivotTable` is common for all types of pivots (spreadsheet or odoo)

Task: 0
