+++ src/index.ts
@@ -432,7 +432,7 @@ export const constants = {
 };
 
 export { PivotRuntimeDefinition } from "./helpers/pivot/pivot_runtime_definition";
-export { SpreadsheetPivotTable } from "./helpers/pivot/spreadsheet_pivot/table_spreadsheet_pivot";
+export { SpreadsheetPivotTable } from "./helpers/pivot/table_spreadsheet_pivot";
 
 export type { EnrichedToken } from "./formulas/composer_tokenizer";
 export type { AST, ASTFuncall } from "./formulas/parser";
