+++ src/types/pivot_runtime.ts
@@ -1,5 +1,5 @@
 import { PivotRuntimeDefinition } from "../helpers/pivot/pivot_runtime_definition";
-import { SpreadsheetPivotTable } from "../helpers/pivot/spreadsheet_pivot/table_spreadsheet_pivot";
+import { SpreadsheetPivotTable } from "../helpers/pivot/table_spreadsheet_pivot";
 import { FunctionResultObject, Maybe } from "./misc";
 import {
   PivotCoreDefinition,
